Implementado Glide (librería de imágenes) y permiso de internet

+++ app/src/main/java/com/example/jimenaproject/ProductsModel.kt
@@ -1,7 +1,7 @@
 package com.example.jimenaproject
 
 data class ProductsModel(
-    val idproduct: Int,
+    val idproduct: Long,
     val productname: String,
     val productdescription: String,
     val photo: String
